// packages/core/src/multichainApiGenerated/models/Activity.ts
export interface Activity {
    /**
     * 
     * @type {ActivityType}
     * @memberof Activity
     */
    activityType: ActivityType;
    /**
     * 
     * @type {ActivityStatus}
     * @memberof Activity
     */
    status: ActivityStatus;
    /**
     * 
     * @type {Date}
     * @memberof Activity
     */
    blockTime: Date;
    /**
     * 
     * @type {number}
     * @memberof Activity
     */
    blockNumber?: number;
    /**
     * 
     * @type {Chain}
     * @memberof Activity
     */
    fromChain: Chain;
    /**
     * 
     * @type {Chain}
     * @memberof Activity
     */
    toChain: Chain;
    /**
     * 
     * @type {string}
     * @memberof Activity
     */
    walletAddress?: string;
    /**
     * 
     * @type {ActivityDirection}
     * @memberof Activity
     */
    direction: ActivityDirection;
    /**
     * 
     * @type {string}
     * @memberof Activity
     */
    fromAddress?: string;
    /**
     * 
     * @type {string}
     * @memberof Activity
     */
    toAddress?: string;
    /**
     * 
     * @type {AssetInfo}
     * @memberof Activity
     */
    outToken?: AssetInfo;
    /**
     * 
     * @type {string}
     * @memberof Activity
     */
    outAmount?: string;
    /**
     * 
     * @type {number}
     * @memberof Activity
     */
    outAmountUsd?: number;
    /**
     * 
     * @type {AssetInfo}
     * @memberof Activity
     */
    inToken?: AssetInfo;
    /**
     * 
     * @type {string}
     * @memberof Activity
     */
    inAmount?: string;
    /**
     * 
     * @type {number}
     * @memberof Activity
     */
    inAmountUsd?: number;
    /**
     * 
     * @type {AssetInfo}
     * @memberof Activity
     */
    feeToken?: AssetInfo;
    /**
     * 
     * @type {string}
     * @memberof Activity
     */
    feeAmount?: string;
    /**
     * 
     * @type {number}
     * @memberof Activity
     */
    feeAmountUsd?: number;
    /**
     * DEX/bridge protocol name (uniswap_v3, stargate, etc.)
     * @type {string}
     * @memberof Activity
     */
    protocol?: string;
    /**
     * On-chain tx identifiers, format: chain:txhash
     * @type {Array<string>}
     * @memberof Activity
     */
    txIds: Array<string>;
    /**
     * 
     * @type {boolean}
     * @memberof Activity
     */
    isRead?: boolean;
    /**
     * Chain-specific extras (memo, input data, etc.)
     * @type {{ [key: string]: any; }}
     * @memberof Activity
     */
    meta?: { [key: string]: any; };
}