// packages/core/src/multichainApiGenerated/models/AssetInfo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tonkeeper Multichain Backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Token metadata (used in both balances and activities)
 * @export
 * @interface AssetInfo
 */
export interface AssetInfo {
    /**
     * Universal asset ID in format chain/network/type/address (e.g. eth/mainnet/coin, ton/mainnet/jetton/EQC...)
     * @type {string}
     * @memberof AssetInfo
     */
    assetId: string;
    /**
     * Asset name
     * @type {string}
     * @memberof AssetInfo
     */
    name: string;
    /**
     * Asset symbol (ticker)
     * @type {string}
     * @memberof AssetInfo
     */
    symbol: string;
    /**
     * Decimals of coin or token
     * @type {number}
     * @memberof AssetInfo
     */
    decimals: number;
    /**
     * Asset image URL
     * @type {string}
     * @memberof AssetInfo
     */
    image: string;
}

/**
 * Check if a given object implements the AssetInfo interface.
 */
export function instanceOfAssetInfo(value: object): value is AssetInfo {
    if ((!('assetId' in (value as Record<string, any>)) && !('asset_id' in (value as Record<string, any>))) || ((value as Record<string, any>)['assetId'] === undefined && (value as Record<string, any>)['asset_id'] === undefined)) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('symbol' in value) || value['symbol'] === undefined) return false;
    if (!('decimals' in value) || value['decimals'] === undefined) return false;
    if (!('image' in value) || value['image'] === undefined) return false;
    return true;
}

export function AssetInfoFromJSON(json: any): AssetInfo {
    return AssetInfoFromJSONTyped(json, false);
}

export function AssetInfoFromJSONTyped(json: any, ignoreDiscriminator: boolean): AssetInfo {
    if (json == null) {
        return json;
    }
    return {
        
        'assetId': json['asset_id'],
        'name': json['name'],
        'symbol': json['symbol'],
        'decimals': json['decimals'],
        'image': json['image'],
    };
}

export function AssetInfoToJSON(json: any): AssetInfo {
    return AssetInfoToJSONTyped(json, false);
}

export function AssetInfoToJSONTyped(value?: AssetInfo | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'asset_id': value['assetId'],
        'name': value['name'],
        'symbol': value['symbol'],
        'decimals': value['decimals'],
        'image': value['image'],
    };
}