// packages/core/src/multichainApiGenerated/models/Chain.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tonkeeper Multichain Backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const Chain = {
    Ton: 'ton',
    Eth: 'eth',
    Base: 'base',
    Btc: 'btc',
    Tron: 'tron',
    Arb: 'arb',
    Bsc: 'bsc'
} as const;
export type Chain = typeof Chain[keyof typeof Chain];


export function instanceOfChain(value: any): boolean {
    for (const key in Chain) {
        if (Object.prototype.hasOwnProperty.call(Chain, key)) {
            if (Chain[key as keyof typeof Chain] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ChainFromJSON(json: any): Chain {
    return ChainFromJSONTyped(json, false);
}

export function ChainFromJSONTyped(json: any, ignoreDiscriminator: boolean): Chain {
    return json as Chain;
}

export function ChainToJSON(value?: Chain | null): any {
    return value as any;
}

export function ChainToJSONTyped(value: any, ignoreDiscriminator: boolean): Chain {
    return value as Chain;
}